Trigger settings
The value  asum_threshold 270 mV is too high for much of the existing LST1 data. We reduce it to 230 mV, together with a reduction the min_photons and min_photoelectrons values settings to be on the safe side (but especially the change in min_photons will increase sim_telarray execution time - @Voutsi, did you check by how much?).